Fix formatting and casing
Not sure if it was the auto-formatting, but the if statements for checking what gender the user has in the database were affected.

For example line 36 looked like:
<option value="male" {% if user.gender="" ="male" %}selected{% endif %}>

And it gave an error of:
jinja2.exceptions.TemplateSyntaxError: expected token 'end of statement block', got '='

So fixed those, and also changed what value is displayed when the condition gets checked to be capitalised.

Apart from that it works great! Good work.