Add Basic Playwright Test
Added a test to check if all the default database posts are displayed.

Meant to be able to run `playwright codegen http://127.0.0.1:5000` and it will generate tests while recording your actions. But can't seem to get it to work. Gives a very corrupted looking view. Found a bit of a workaround as the Node.js version of playwright doesn't have this issue. But then need to convert the Typescript into Python.

Can run the test by running the command `pytest` in either the home or src directories. Have added an updated dependency list in `testreqs.txt` but I think I can slim it down a bit.