Add 5 basic Playwright tests
Added 5 tests that can be run by navigating to the src folder and running " pytest " , can run with --headed if you want to see it as it tests.

Currently have tests for:
- Basic login
- Password change (This branch was made before validation though, so will remake it)
- Removing notifications
- Creating a post
- Checking the default posts on browse

Will definitely need to PR just to get up to date with the new database generation and new features.